python_ai_service/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
numpy==1.26.0
pandas==2.1.1
python-multipart==0.0.6
pytorch==2.1.0
transformers==4.34.1
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.3
matplotlib==3.8.0
scikit-learn==1.3.1
openai==1.2.0
reportlab==4.0.7
python-dotenv==1.0.0
spacy==3.7.2
requests==2.31.0
